+++ tests/org.preesm.tests.framework/src/org/preesm/commons/test/WorkspaceUtilsTest.java
@@ -1,7 +1,7 @@
 /**
- * Copyright or © or Copr. IETR/INSA - Rennes (2017 - 2019) :
+ * Copyright or © or Copr. IETR/INSA - Rennes (2017 - 2018) :
  *
- * Antoine Morvan <antoine.morvan@insa-rennes.fr> (2017 - 2019)
+ * Antoine Morvan <antoine.morvan@insa-rennes.fr> (2017 - 2018)
  *
  * This software is a computer program whose purpose is to help prototyping
  * parallel applications using dataflow formalism.
@@ -32,30 +32,32 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL license and that you accept its terms.
  */
-package org.preesm.ui.workflow.command;
+package org.preesm.commons.test;
 
-import java.io.File;
-import java.io.IOException;
-import org.eclipse.core.expressions.PropertyTester;
-import org.eclipse.core.resources.IFile;
-import org.preesm.commons.exceptions.PreesmRuntimeException;
-import org.preesm.workflow.converter.WorkflowConverter;
+import org.junit.Assert;
+import org.junit.Test;
+import org.preesm.commons.files.WorkspaceUtils;
 
+// TODO: Auto-generated Javadoc
 /**
+ * The Class ContainersManagerTest.
  */
-public class OldWorkflowPropertyTester extends PropertyTester {
+public class WorkspaceUtilsTest {
 
-  @Override
-  public boolean test(final Object receiver, final String property, final Object[] args, final Object expectedValue) {
-
-    final IFile file = (IFile) receiver;
-    final File file2 = file.getLocation().toFile();
+  /**
+   * Test project exists.
+   */
+  @Test
+  public void testProjectExists() {
     try {
-      final boolean newWorkflow = WorkflowConverter.isNewWorkflow(file2);
-      return !newWorkflow;
-    } catch (final IOException e) {
-      throw new PreesmRuntimeException("Could not access file", e);
+      final boolean projectExists = WorkspaceUtils.projectExists("toto");
+      Assert.assertFalse(projectExists);
+      System.out.println("done");
+    } catch (final IllegalStateException e) {
+      System.out.println("workspace not found");
+      return;
     }
+
   }
 
 }